Create GPAC-Master, recovering programs and windows?
Howdy,
In an attempt to fix an "unfinished" clip that no editing program or player was reading, I downloaded GPAC-Master.
I tried according to some instructions how to do it, but they didn't work. I do not know what was in my head, but I gave him "Clean" with "Run as administrator" and I let him clean, hoping that it will only touch on the temporary files resulting from numerous video edits.
Now I notice that I no longer have a functional program, I tried to recover the windows, and the restore points deleted them, everything .. Windows works, but I am afraid to restart that it will no longer work. and he.
What else could I do in this case?
Thanks in advance for the help!